requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==4.6.1
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.2.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
graphene-django==3.1.5
channels==4.0.0
daphne==4.0.0
django-filter==24.1
django-redis==5.4.0
django-storages==1.14.2
boto3==1.28.62
gunicorn==21.2.0
whitenoise==6.5.0
sqlparse==0.2.4
djangorestframework-simplejwt==5.3.1 